SourceEngine: Change `VPK::Entry` to own it's data, instead of the `VPK`
Instead of asking the API user to read an entire archive for us into a
`ByteBuffer` (which is ~100MB for every archive), we now ask it for a
`Core::Stream::SeekableStream`, which allows us to seek around and only
read the necessary parts of the file. This allows `VPK::Entry` to own
it's data instead, and stops us from making 100MB allocations for each
archive, which would be contiguous. Although, we are making more overall
allocations (one for each entry, instead of one for each archive).

This doesn't really reduce overall memory useage, as the same amount of
memory is still going to be used as the entire archive is likely to be
mapped to an `VPK::Entry`.

Memory usage is still a large problem with this implementation -- we
_should_ defer loading the data of an entry from an archive to only upon
request, and optionally cache it?