SourceEngine: Defer reading of `VPK::Entry` data until requested
We shouldn't read all of the archives (and entry data) all at once,
because it'll often be many gigabytes in size, which would quickly
exhaust us. Instead, have the `VPK` own the
`Core::Stream::SeekableStream`s that the user API has given us, and give
each `VPK::Entry` a pointer to that stream. This allows each entry to
read it's own data whenever it needs to, and not immediately when
parsing the VPK file.