Everywhere: Fix compilation issues after change to `Core::Stream::read`
`Core::Stream::read` now returns a `Bytes` with the same pointer as the
passed `Bytes`, but the length is changed to the actual amount of bytes
read instead.

`Core::File::ensure_parent_directories` was also removed, in favor of a
somewhat-replacement `Core::Directory::create`. We have to calculate the
parent directory from a file ourself, but it completely removes the need
to do any `realpath` or dumb pseudo `realpath` dumbness.